+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * Obligations and duties may now have notes set, and the character importer correctly populates this field
   * Defense dice are now added if an actor is targeted (this is configurable per-client)
   * Items imported via OggDude importer now include the categories from the source dataset (e.g. the Holdout Blaster is tagged with "Blaster Pistol", "Holdout blaster", "Blaster", "Pistol", "Ranged")
+  * Abilities created during the SWA import process are now created as "Ability" items instead of populating in the biography section
 * Fixes:
   * Sending a signature ability to chat now includes purchased upgrades
   * Granting XP to the entire group now updates the XP logs